@@ -4,65 +4,67 @@
   "homepage": ".",
   "private": true,
   "devDependencies": {
-    "custom-react-scripts": "^0.2.0",
     "dotenv": "^5.0.1",
     "eslint-config-prettier": "^2.9.0",
     "eslint-plugin-prettier": "^2.6.0",
-    "gulp-clean": "^0.3.2",
-    "gulp-template": "^4.0.0",
-    "prettier": "^1.12.1",
+    "node-sass": "^4.12.0",
+    "prettier": "1.12.1",
+    "react-scripts": "2.1.8",
     "run-sequence": "^1.2.2"
   },
   "dependencies": {
     "@mapbox/geojson-area": "^0.2.2",
+    "@mapbox/togeojson": "^0.16.0",
+    "@turf/area": "^6.0.1",
+    "@turf/invariant": "^6.1.2",
+    "@turf/length": "^6.0.2",
     "axios": "^0.17.1",
     "axios-extensions": "^2.0.0",
     "client-oauth2": "^4.1.0",
+    "d3-queue": "^3.0.7",
     "del": "^2.2.2",
     "dlv": "^1.1.0",
     "dragula": "^3.7.1",
+    "eo-components": "0.0.23-rc.4",
     "file-saver": "^1.3.3",
     "font-awesome": "^4.6.3",
     "geojson-utils": "^1.1.0",
     "gifshot": "https://github.com/guss84/gifshot/archive/master.tar.gz",
-    "gulp": "^3.9.1",
-    "gulp-template": "^4.0.0",
     "js-md5": "^0.4.1",
     "jszip": "^3.1.5",
     "jszip-utils": "^0.0.2",
     "jwt-decode": "^2.2.0",
-    "leaflet": "1.0.2",
-    "leaflet.pm": "^0.23.1",
+    "leaflet": "^1.3.1",
+    "leaflet.pm": "^0.25.0",
     "lodash": "^4.15.0",
+    "mapbox-gl": "^0.49.0",
     "moment": "^2.21.0",
     "nprogress": "^0.2.0",
     "proj4": "2.3.15",
     "prop-types": "^15.6.1",
-    "rc-slider": "^4.0.1",
-    "react": "^15.3.1",
+    "rc-slider": "^8.6.1",
+    "react": "^16.4.0",
     "react-addons-shallow-compare": "^15.4.2",
     "react-alert": "2.4.0",
     "react-awesome-modal": "^2.0.3",
     "react-checkbox-group": "^3.0.1",
-    "react-codemirror": "^0.2.6",
+    "react-codemirror": "^1.0.0",
     "react-day-picker": "^7.1.9",
-    "react-dom": "^15.3.1",
+    "react-dom": "^16.4.0",
     "react-dragula": "^1.1.16",
     "react-dropzone": "^4.1.2",
-    "react-geosuggest": "^1.25.0",
+    "react-joyride": "2.0.0-14",
     "react-onclickoutside": "^5.5.0",
     "react-radio-group": "^3.0.1",
     "react-redux": "^4.4.5",
-    "react-sidebar": "^2.3.2",
-    "react-skylight": "^0.5.0",
-    "react-tabs": "^0.8.0",
-    "react-toggle": "^2.2.1",
-    "react-toolbox": "^2.0.0-beta.12",
+    "react-toggle": "^4.0.2",
+    "react-tooltip-lite": "^1.5.2",
     "redux": "^3.6.0",
     "redux-observable": "^0.9.1",
     "rodal": "^1.3.2",
     "rxjs": "5.0.3",
-    "semiotic": "^1.8.3",
+    "semiotic": "^1.11.1",
+    "sentinelhub-js": "git+https://gitext.sinergise.com/team-6/sentinelhub-js.git#1639bc649434429fc61abbf65b0182ec195be22f",
     "source-map-explorer": "^1.5.0",
     "styled-components": "^2.1.2",
     "terraformer": "^1.0.8",
@@ -74,18 +76,30 @@
   },
   "scripts": {
     "analyze": "source-map-explorer build/static/js/main.*",
-    "prettier": "prettier --write 'src/**/*.js'",
-    "start": "node -r dotenv/config ./node_modules/.bin/react-scripts start dotenv_config_path=config/development.env",
+    "prettier": "prettier --write \"src/**/*.js\"",
+    "start": "react-scripts start",
     "build": "node -r dotenv/config ./node_modules/.bin/react-scripts build dotenv_config_path=config/${BUILD_ENV}.env  ",
     "build:testing": "BUILD_ENV=testing npm run build",
     "build:staging": "BUILD_ENV=staging npm run build",
     "build:production": "BUILD_ENV=production npm run build",
     "lint": "eslint --max-warnings 0 --ext .js src/",
-    "test": "react-app-env test",
+    "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
     "see": "node -e 'console.log(require(\"process\").env)'"
   },
   "eslintConfig": {
     "extends": "./node_modules/react-scripts/config/eslint.js"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
