@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
 import { uniquePinId } from '../utils/utils';
 import Store from '../store';
-import { connect } from 'react-redux';
+import App from '../App';
 
 class AddPin extends Component {
   savePin = () => {
@@ -19,20 +20,28 @@ class AddPin extends Component {
       lat: pin ? pin.lat : lat,
       lng: pin ? pin.lng : lng,
       _id: uniquePinId(),
+      pinTitle: '',
     };
-    Store.addPinResult(savePin);
+    Store.addPins([savePin]);
     Store.setTabIndex(3);
   };
 
   render() {
+    const { loggedIn } = this.props;
+    const title = `Pin to your favourite items${
+      loggedIn ? '' : ' (you need to log in to use this function)'
+    }`;
     return (
       <a
         className="addToPin"
         onClick={() => {
-          // this.props.onAddToPin && this.props.onAddToPin();
+          if (!loggedIn) {
+            App.displayErrorMessage(title);
+            return;
+          }
           this.savePin();
         }}
-        title="Pin to your favourite items"
+        title={title}
       >
         <i className="fa fa-thumb-tack" />
       </a>
@@ -40,4 +49,12 @@ class AddPin extends Component {
   }
 }
 
-export default connect(s => s)(AddPin);
+const mapStoreToProps = store => ({
+  lat: store.lat,
+  lng: store.lng,
+  zoom: store.zoom,
+  selectedResult: store.selectedResult,
+  presets: store.presets,
+  loggedIn: !!store.user,
+});
+export default connect(mapStoreToProps)(AddPin);
