@@ -2,24 +2,27 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import dragula from 'react-dragula';
 import 'react-dragula/dist/dragula.min.css';
-import NotificationPanel from './NotificationPanel';
+import { NotificationPanel } from 'eo-components';
 import Store from '../store';
 import Pin from './Pin';
 import Button from './Button';
 import { RadioGroup, Radio } from 'react-radio-group';
 import Rodal from 'rodal';
+import PinTools from './PinTools';
+import UploadPinsJsonFile from './PinTools/UploadPinsJsonFile';
 
 class PinPanel extends React.Component {
+  constructor() {
+    super();
+    this.state = { uploadDialog: false };
+  }
+
   dragulaDecorator = componentBackingInstance => {
     if (componentBackingInstance) {
       dragula([componentBackingInstance], {
         moves: (el, container, handle) => {
-          if (this.props.isCompare) {
-            if (handle.classList.contains('rc-slider-handle')) {
-              return false;
-            } else {
-              return true;
-            }
+          if (handle.classList.contains('pin-dragHandler') || handle.classList.contains('fa-ellipsis-v')) {
+            return true;
           }
         },
       }).on('drop', (el, target, source, sibling) => {
@@ -42,6 +45,7 @@ class PinPanel extends React.Component {
         width={400}
         height={150}
         onClose={() => this.closeDialog(modalDialogId)}
+        closeOnEsc={true}
       >
         <h3>Delete pins</h3>
         <b>WARNING:</b> You're about to delete all pins. Do you wish to continue? <br />
@@ -72,22 +76,58 @@ class PinPanel extends React.Component {
 
   changePinOrder = (oldIndex, newIndex) => {
     this.props.pinOrderChange(oldIndex, newIndex);
-    Store.changePinOrder(oldIndex, newIndex);
+    Store.changePinOrder(oldIndex, newIndex, this.props.readOnly);
+  };
+
+  openUploadPinsJsonFileDialog = () => {
+    this.setState({ uploadDialog: true });
+  };
+
+  onUpload = (uploadedPins, keepExisting) => {
+    if (!keepExisting) {
+      this.props.onClearPins();
+    }
+    let filteredUpload = uploadedPins.filter(upin => !this.props.items.find(pin => pin._id === upin._id));
+    filteredUpload.reverse();
+    Store.addPins(filteredUpload);
+    this.setState({ uploadDialog: false });
+    Store.setTabIndex(3);
+  };
+
+  safeOnCompare = () => {
+    if (this.props.items.length > 0) {
+      this.props.onCompare();
+    }
   };
 
   render() {
-    let { items, isCompare, isOpacity, onToggleCompareMode, onRemove, onCompare } = this.props;
+    let { items, isCompare, isOpacity, onToggleCompareMode, onRemove, loggedIn } = this.props;
+    const noPinMsg = `No pins. ${
+      this.props.readOnly
+        ? 'This theme has no pins stored'
+        : 'Go to the Visualization tab to save a pin or upload a JSON file with saved pins'
+    }.`;
     return (
       <div className={`pinPanel ${!isCompare && 'normalMode'}`}>
-        {items.length === 0 ? (
-          <NotificationPanel type="info" msg="No pins. Find your scene and pin it to save it for later." />
+        {!loggedIn ? (
+          <NotificationPanel type="info" msg="Please log in to access your pins." />
         ) : (
           <div>
             <div className="comparisonHeader">
-              <a style={{ float: 'right' }} onClick={this.confirmDeleteAllPins}>
-                <i className="fa fa-trash" />Clear pins
-              </a>
-              <a onClick={onCompare}>
+              <PinTools
+                pins={items}
+                confirmDeleteAllPins={this.confirmDeleteAllPins}
+                openUploadPinsJsonFileDialog={this.openUploadPinsJsonFileDialog}
+              />
+
+              {this.state.uploadDialog && (
+                <UploadPinsJsonFile
+                  onUpload={this.onUpload}
+                  onClose={() => this.setState({ uploadDialog: false })}
+                />
+              )}
+
+              <a onClick={this.safeOnCompare} className={items.length === 0 ? 'disabled' : ''}>
                 <i className={`fa fa-${isCompare ? 'check-circle-o' : 'exchange'}`} />
                 {isCompare ? 'Finish comparison' : 'Compare'}
               </a>
@@ -110,26 +150,32 @@ class PinPanel extends React.Component {
                 </div>
               )}
             </div>
-            <div ref={this.dragulaDecorator}>
-              {items.map((item, i) => {
-                // if we add pin from results, we have different structure, otherwise we have all the data
-                return (
-                  <Pin
-                    ref={'pin' + i}
-                    range={[0, 1]}
-                    item={item}
-                    index={i}
-                    key={item._id}
-                    isOpacity={isOpacity}
-                    onRemove={onRemove}
-                    onZoomToPin={this.props.onZoomToPin}
-                    isCompare={isCompare}
-                    onPinClick={() => this.props.onPinClick(item, i, false)}
-                    onOpacityChange={e => this.props.onOpacityChange(e, i)}
-                  />
-                );
-              })}
-            </div>
+
+            {items.length === 0 ? (
+              <NotificationPanel type="info" msg={noPinMsg} />
+            ) : (
+              <div ref={this.dragulaDecorator}>
+                {items.map((item, i) => {
+                  // if we add pin from results, we have different structure, otherwise we have all the data
+                  return (
+                    <Pin
+                      ref={'pin' + i}
+                      range={[0, 1]}
+                      item={item}
+                      index={i}
+                      key={item._id}
+                      isOpacity={isOpacity}
+                      onRemove={onRemove}
+                      onZoomToPin={this.props.onZoomToPin}
+                      isCompare={isCompare}
+                      onPinClick={this.props.onPinClick}
+                      onOpacityChange={e => this.props.onOpacityChange(e, i)}
+                      readOnly={this.props.readOnly}
+                    />
+                  );
+                })}
+              </div>
+            )}
           </div>
         )}
       </div>
@@ -139,6 +185,7 @@ class PinPanel extends React.Component {
 
 PinPanel.propTypes = {
   zoom: PropTypes.number,
+  loggedIn: PropTypes.bool,
   isCompare: PropTypes.bool,
   isOpacity: PropTypes.bool,
   items: PropTypes.array,
