@@ -1,56 +1,142 @@
 import React from 'react';
 import geo_area from '@mapbox/geojson-area';
+import moment from 'moment';
+import Rodal from 'rodal';
 import Store from '../../store';
 import App from '../../App';
+import MeasureIcon from './MeasureIcon';
 import { isCustomPreset } from '../../utils/utils';
 import { getFisShadowLayer } from '../FIS';
+import ImageDownloadPanel from '../ImageDownload/ImageDownloadPanel';
+import Timelapse from '../Timelapse';
 
-const NOT_LOGGED_IN_ERROR_MESSAGE = 'You need to log in to use this function';
+import Tutorial from '../Tutorial/Tutorial';
 
-export class DrawMarker extends React.Component {
+const NOT_LOGGED_IN_ERROR_MESSAGE = 'you need to log in to use this function';
+const LAYER_NOT_SELECTED_ERROR_MESSAGE = 'please select a layer';
+const COMPARE_MODE_ERROR_MESSAGE = 'downloading image in compare mode is not possible';
+const DATASOURCE_NOT_SUPPORTED_ERROR_MESSAGE = 'this datasource is not supported';
+
+const PrettyDistance = ({ distance }) => {
+  const divided = distance / 1000;
+  if (divided >= 1) {
+    return <span>{divided.toFixed(2)} km</span>;
+  } else {
+    return <span>{distance.toFixed()} m</span>;
+  }
+};
+
+const PrettyArea = ({ area }) => {
+  const areaKM = area / 1000000;
+  return (
+    <span>
+      {areaKM.toFixed(3)} km<sup>2</sup>
+    </span>
+  );
+};
+export class MeasurePanelButton extends React.Component {
+  showMeasureInfo = () => (
+    <span className="aoiCords">
+      {this.props.distance && (
+        <div className="measure-text">
+          <PrettyDistance distance={this.props.distance} />
+        </div>
+      )}
+      {this.props.area !== 0 && this.props.area ? (
+        <div className="measure-text">
+          <PrettyArea area={this.props.area} />
+        </div>
+      ) : null}
+      <span>
+        <a onClick={this.props.removeMeasurement} title="Remove measurement">
+          <i className={`fa fa-close`} />
+        </a>
+      </span>
+    </span>
+  );
+
+  renderMeasureIcon = () => {
+    const errMsg = this.props.isLoggedIn ? null : NOT_LOGGED_IN_ERROR_MESSAGE;
+    const isEnabled = errMsg === null;
+    const title = `Measure distances ${errMsg ? ` (${errMsg})` : ''}`;
+    return (
+      <a
+        className={`drawGeometry ${this.props.isLoggedIn ? '' : 'disabled'}`}
+        onClick={ev => {
+          if (!isEnabled) {
+            App.displayErrorMessage(title);
+            return;
+          }
+          this.props.toggleMeasure();
+        }}
+        title={title}
+      >
+        <i>
+          <MeasureIcon />
+        </i>
+      </a>
+    );
+  };
+
+  render() {
+    return (
+      <div className="measurePanel panelButton floatItem">
+        {this.props.hasMeasurement && this.showMeasureInfo()}
+        {this.renderMeasureIcon()}
+      </div>
+    );
+  }
+}
+
+export class POIPanelButton extends React.Component {
   state = {
     showOptions: false,
   };
-  renderMarkerIcon = () => (
-    <span
-      onClick={ev => {
-        if (this.props.disabled) {
-          App.displayErrorMessage(NOT_LOGGED_IN_ERROR_MESSAGE);
-          return;
-        }
-        this.props.drawMarker();
-      }}
-    >
-      <a
-        className={this.props.disabled ? 'drawGeometry disabled' : 'drawGeometry'}
-        title={!this.props.disabled ? 'Draw point of interest' : NOT_LOGGED_IN_ERROR_MESSAGE}
+
+  renderMarkerIcon = () => {
+    const errMsg = this.props.disabled ? NOT_LOGGED_IN_ERROR_MESSAGE : null;
+    const isEnabled = errMsg === null;
+    const title = `Mark point of interest${errMsg ? ` (${errMsg})` : ''}`;
+    return (
+      <span
+        onClick={ev => {
+          if (!isEnabled) {
+            App.displayErrorMessage(title);
+            return;
+          }
+          this.props.drawMarker();
+        }}
+        title={title}
       >
-        <i className="fa fa-map-marker" />
-      </a>
-    </span>
-  );
+        <a className={`drawGeometry ${this.props.disabled ? 'disabled' : ''}`}>
+          <i className="fa fa-map-marker" />
+        </a>
+      </span>
+    );
+  };
+
   renderMarkerInfo = () => (
     <span>
       <a onClick={() => this.props.centerOnFeature('poiLayer')} title="Center map on feature">
         <i className={`fa fa-crosshairs`} />
       </a>
-      {this.props.poi &&
-        Store.current.selectedResult && (
-          <FisChartLink
-            aoiOrPoi={'poi'}
-            selectedResult={this.props.selectedResult}
-            openFisPopup={this.props.openFisPopup}
-          />
-        )}
+      {this.props.poi && (
+        <FisChartLink
+          aoiOrPoi={'poi'}
+          selectedResult={this.props.selectedResult}
+          openFisPopup={this.props.openFisPopup}
+        />
+      )}
       <a onClick={this.props.deleteMarker} title={'Remove geometry'}>
         <i className={`fa fa-close`} />
       </a>
     </span>
   );
+
   render() {
     const { poi } = this.props;
     return (
-      <div className="poiPanel floatItem" title="Area of interest">
+      <div className="poiPanel panelButton floatItem" title="Area of interest">
         {poi && !this.props.disabled && this.renderMarkerInfo()}
         {this.renderMarkerIcon()}
       </div>
@@ -58,7 +144,7 @@ export class DrawMarker extends React.Component {
   }
 }
 
-export class DrawArea extends React.Component {
+export class AOIPanelButton extends React.Component {
   state = {
     showOptions: false,
   };
@@ -70,6 +156,7 @@ export class DrawArea extends React.Component {
     }
     this.setState({ showOptions: true });
   };
+
   hideOptions = () => {
     this.hideOptionsTimeout = setTimeout(() => {
       this.setState({ showOptions: false });
@@ -78,10 +165,11 @@ export class DrawArea extends React.Component {
 
   renderOptionButtons = () => (
     <div className="aoiCords">
-      <OpenKmkDialogButton handleClick={this.props.openUploadKMLDialog} />
+      <OpenUploadDataDialogButton handleClick={this.props.openUploadGeoFileDialog} />
       <DrawButton geomType={'polygon'} />
     </div>
   );
+
   renderAioInfo = () => (
     <span className="aoiCords">
       <span className="area-text">
@@ -98,42 +186,43 @@ export class DrawArea extends React.Component {
         <a onClick={() => this.props.centerOnFeature('aoiLayer')} title="Center map on feature">
           <i className={`fa fa-crosshairs`} />
         </a>
-        {this.props.aoiBounds &&
-          Store.current.selectedResult && (
-            <FisChartLink
-              selectedResult={this.props.selectedResult}
-              openFisPopup={this.props.openFisPopup}
-              aoiOrPoi={'aoi'}
-            />
-          )}
+        {this.props.aoiBounds && (
+          <FisChartLink
+            selectedResult={this.props.selectedResult}
+            openFisPopup={this.props.openFisPopup}
+            aoiOrPoi={'aoi'}
+          />
+        )}
         <a onClick={this.props.resetAoi} title={this.props.isAoiClip ? 'Cancel edit.' : 'Remove geometry'}>
           <i className={`fa fa-close`} />
         </a>
       </span>
     </span>
   );
+
   render() {
     const { aoiBounds, isAoiClip } = this.props;
     const { showOptions } = this.state;
     const doWeHaveAOI = aoiBounds || isAoiClip;
     const showOptionsMenu = !doWeHaveAOI && showOptions;
+    const errMsg = this.props.disabled ? NOT_LOGGED_IN_ERROR_MESSAGE : null;
+    const isEnabled = errMsg === null;
+    const title = `Draw area of interest${errMsg ? ` (${errMsg})` : ''}`;
     return (
       <div
-        className="aoiPanel floatItem"
-        onMouseEnter={!this.props.disabled && this.showOptions}
-        onMouseLeave={!this.props.disabled && this.hideOptions}
+        className="aoiPanel panelButton floatItem"
+        onMouseEnter={!this.props.disabled ? this.showOptions : null}
+        onMouseLeave={!this.props.disabled ? this.hideOptions : null}
         onClick={ev => {
-          if (this.props.disabled) {
-            App.displayErrorMessage(NOT_LOGGED_IN_ERROR_MESSAGE);
+          if (!isEnabled) {
+            App.displayErrorMessage(title);
           }
         }}
+        title={title}
       >
         {showOptionsMenu && this.renderOptionButtons()}
         {doWeHaveAOI && !this.props.disabled && this.renderAioInfo()}
-        <a
-          className={this.props.disabled ? 'drawGeometry disabled' : 'drawGeometry'}
-          title={!this.props.disabled ? 'Draw area of interest' : NOT_LOGGED_IN_ERROR_MESSAGE}
-        >
+        <a className={`drawGeometry ${this.props.disabled ? 'disabled' : ''}`}>
           <i>
             <PolygonSvgIcon />
           </i>
@@ -152,8 +241,8 @@ const DrawButton = ({ geomType }) => (
   </a>
 );
 
-const OpenKmkDialogButton = ({ handleClick }) => (
-  <a title="Upload KML/KMZ file" onClick={handleClick}>
+const OpenUploadDataDialogButton = ({ handleClick }) => (
+  <a title="Upload data" onClick={handleClick}>
     <i className="fa fa-upload" />
   </a>
 );
@@ -173,10 +262,12 @@ const PolygonSvgIcon = ({ fillColor }) => (
 );
 
 const FisChartLink = props => {
-  const isCustomLayer = isCustomPreset(props.selectedResult.preset);
-  const isShadowLayerAvailable = !!getFisShadowLayer(props.selectedResult.name, props.selectedResult.preset);
-  const isFisAvailableOnDatasource = props.selectedResult && props.selectedResult.baseUrls.FIS;
-  if (isFisAvailableOnDatasource && (isShadowLayerAvailable || isCustomLayer)) {
+  const isSelectedResult = !!props.selectedResult;
+  const isCustomLayer = props.selectedResult && isCustomPreset(props.selectedResult.preset);
+  const isShadowLayerAvailable =
+    props.selectedResult && !!getFisShadowLayer(props.selectedResult.name, props.selectedResult.preset);
+  const isFisAvailableOnDatasource = !!(props.selectedResult && props.selectedResult.baseUrls.FIS);
+  if (isSelectedResult && isFisAvailableOnDatasource && (isShadowLayerAvailable || isCustomLayer)) {
     return (
       <a
         onClick={() => props.openFisPopup(props.aoiOrPoi)}
@@ -186,16 +277,23 @@ const FisChartLink = props => {
       </a>
     );
   } else {
+    const errorMessage = `Statistical Info / Feature Info Service chart - ${
+      !isSelectedResult
+        ? 'please select a layer'
+        : !isFisAvailableOnDatasource
+          ? 'not available for ' + props.selectedResult.name
+          : `not available for "${
+              Store.current.presets[props.selectedResult.name].find(l => l.id === props.selectedResult.preset)
+                .name
+            }" (layer with value is not set up)`
+    }`;
     return (
       <a
         onClick={e => {
           e.preventDefault();
+          App.displayErrorMessage(errorMessage);
         }}
-        title={`Statistical Info / Feature Info Service chart - ${
-          !isFisAvailableOnDatasource
-            ? 'not available for ' + props.selectedResult.name
-            : `not available for ${props.selectedResult.preset} (layer with value is not set up)`
-        }`}
+        title={errorMessage}
         className="disabled"
       >
         <i className={`fa fa-bar-chart`} />
@@ -203,3 +301,119 @@ const FisChartLink = props => {
     );
   }
 };
+
+export class DownloadPanelButton extends React.Component {
+  triggerImageDownload = () => {
+    const modalDialogId = 'analytical-download';
+    Store.addModalDialog(
+      modalDialogId,
+      <Rodal
+        animation="slideUp"
+        customStyles={{
+          height: 'auto',
+          maxHeight: '80vh',
+          bottom: 'auto',
+          width: '750px',
+          top: '5vh',
+          overflow: 'auto',
+        }}
+        visible
+        onClose={() => Store.removeModalDialog(modalDialogId)}
+        closeOnEsc={true}
+      >
+        <ImageDownloadPanel originalMap={this.props.originalMap} />
+      </Rodal>,
+    );
+  };
+
+  render() {
+    const isLayerSelected = !!Store.current.selectedResult;
+    const isCompareMode = Store.current.compareMode;
+
+    const errMsg = isCompareMode
+      ? COMPARE_MODE_ERROR_MESSAGE
+      : !isLayerSelected
+        ? LAYER_NOT_SELECTED_ERROR_MESSAGE
+        : !Store.current.selectedResult.baseUrls.WMS
+          ? DATASOURCE_NOT_SUPPORTED_ERROR_MESSAGE
+          : null;
+    const isEnabled = errMsg === null;
+    const title = `Download image${errMsg ? ` (${errMsg})` : ''}`;
+    return (
+      <div
+        className="screenshotPanelButton panelButton floatItem"
+        title={title}
+        onClick={ev => {
+          if (!isEnabled) {
+            App.displayErrorMessage(title);
+            return;
+          }
+          this.triggerImageDownload();
+        }}
+      >
+        <a className={`drawGeometry ${isEnabled ? '' : 'disabled'}`}>
+          <i className="fa fa-file-image-o" />
+        </a>
+      </div>
+    );
+  }
+}
+
+export class TimelapsePanelButton extends React.Component {
+  showTimelapse = () => {
+    const modalDialogId = 'timelapse';
+    const datasourceName = Store.current.selectedResult.datasource;
+    Store.addModalDialog(
+      modalDialogId,
+      <Timelapse
+        datasource={datasourceName}
+        startDate={moment(Store.current.selectedResult.time)}
+        captionPrefix={datasourceName || 'Sentinel-2'}
+        onClose={() => Store.removeModalDialog(modalDialogId)}
+      />,
+    );
+  };
+
+  render() {
+    const isLayerSelected = !!Store.current.selectedResult;
+    const isCompareMode = Store.current.compareMode;
+    const isLoggedIn = !!Store.current.user;
+    const isTimelapseSupported =
+      isLayerSelected && Store.current.selectedResult.getDates && Store.current.selectedResult.baseUrls.WMS;
+
+    const errMsg = isCompareMode
+      ? COMPARE_MODE_ERROR_MESSAGE
+      : !isLoggedIn
+        ? NOT_LOGGED_IN_ERROR_MESSAGE
+        : !isLayerSelected
+          ? LAYER_NOT_SELECTED_ERROR_MESSAGE
+          : !isTimelapseSupported
+            ? DATASOURCE_NOT_SUPPORTED_ERROR_MESSAGE
+            : null;
+    const isEnabled = errMsg === null;
+    const title = `Create timelapse animation${errMsg ? ` (${errMsg})` : ''}`;
+    return (
+      <div
+        className="timelapsePanelButton panelButton floatItem"
+        title={title}
+        onClick={ev => {
+          if (!isEnabled) {
+            App.displayErrorMessage(title);
+            return;
+          }
+          this.showTimelapse();
+        }}
+      >
+        <a className={`drawGeometry ${isEnabled ? '' : 'disabled'}`}>
+          <i className="fa fa-film" />
+        </a>
+      </div>
+    );
+  }
+}
+
+export class InfoPanelButton extends React.Component {
+  render() {
+    return <Tutorial />;
+  }
+}
