@@ -4,15 +4,12 @@ import L from 'leaflet';
 import cloneDeep from 'lodash/cloneDeep';
 import isEqual from 'lodash/isEqual';
 import NProgress from 'nprogress';
-import {
-  getCoordsFromBounds,
-  roundDegrees,
-  calcBboxFromXY,
-  bboxToPolygon,
-  getMapDOMSize,
-} from '../../utils/coords';
-import { createMapLayer, evalSourcesMap, isCustomPreset } from '../../utils/utils';
+import { roundDegrees, calcBboxFromXY, bboxToPolygon, getMapDOMSize } from '../../utils/coords';
+import { createMapLayer, isCustomPreset, evalscriptoverridesToString } from '../../utils/utils';
+import { evalSourcesMap } from '../../store/config';
 import '../ext/leaflet-clip-wms-layer';
+import '../ext/leaflet-mapbox-gl';
+import '../ext/leaflet-ruler';
 import Store from '../../store';
 import get from 'dlv';
 import { connect } from 'react-redux';
@@ -20,11 +17,26 @@ import 'nprogress/nprogress.css';
 import 'leaflet.pm';
 import 'leaflet.pm/dist/leaflet.pm.css';
 import gju from 'geojson-utils';
-import UploadKML from '../UploadKML';
+import UploadGeoFile from '../UploadGeoFile';
 import FIS from '../FIS';
-import { DrawArea, DrawMarker } from './DrawVectors';
+import '../ext/leaflet-ruler';
+import '../ext/color-labels';
+
+import {
+  AOIPanelButton,
+  POIPanelButton,
+  DownloadPanelButton,
+  TimelapsePanelButton,
+  MeasurePanelButton,
+  InfoPanelButton,
+} from './DrawVectors';
 import './RootMap.scss';
 import pin from './pin.png';
+import { DEFAULT_POLY_STYLE, HIGHLIGHT_POLY_STYLE } from '../../store/config';
+import roadStyle from './roadStyle.json';
+import boundariesStyle from './boundariesStyle.json';
+import 'mapbox-gl/dist/mapbox-gl.css';
+
 let DefaultIcon = L.icon({
   iconUrl: pin,
   iconAnchor: [13, 40],
@@ -32,13 +44,16 @@ let DefaultIcon = L.icon({
 
 L.Marker.prototype.options.icon = DefaultIcon;
 
+export const DEFAULT_RESULTS_GROUP = 'default';
+
 const COMPARE_LAYER_PANE = 'compareLayer';
 const ACTIVE_LAYER_PANE = 'activeLayer';
 const EDIT_LAYER_PANE = 'editingLayerPane';
 const LABEL_LAYER_PANE = 'labelLayer';
 const MARKER_LAYER_PANE = 'markerPane';
-const OVERLAY_LAYER_PANE = 'overlayPane';
-
+const ROAD_LAYER_PANE = 'roadLayerPane';
+const ADMIN_LAYER_PANE = 'adminLayerPane';
+const MAPBOX_TOKEN = `pk.eyJ1IjoiZ2FleHRyYSIsImEiOiJjajE3cHNrMnMwMDJtMnFuNXoydjI2ODQxIn0.05mV754pUS0SjiThIUx96A`;
 class RootMap extends React.Component {
   constructor(props) {
     super(props);
@@ -50,6 +65,9 @@ class RootMap extends React.Component {
       location: [],
       instances: {},
       uploadDialog: false,
+      hasMeasurement: false,
+      measureDistance: null,
+      measureArea: null,
     };
     this.aoiLayer = null;
     this.mainMap = null;
@@ -60,8 +78,7 @@ class RootMap extends React.Component {
   }
 
   componentDidMount() {
-    const { mapId, lat, lng, zoom } = this.props;
-
+    const { mapId, lat, lng, zoom, mapMaxBounds } = this.props;
     this.progress = NProgress.configure({
       showSpinner: false,
       parent: `#${mapId}`,
@@ -75,6 +92,7 @@ class RootMap extends React.Component {
         // http://{s}.tiles.wmflabs.org/bw-mapnik/{z}/{x}/{y}.png', {
         attribution: 'Carto © CC BY 3.0, OpenStreetMap © ODbL',
         name: 'Carto Light',
+        print: false,
       },
     );
     const cartoVoyagerNoLabels = L.tileLayer(
@@ -82,50 +100,57 @@ class RootMap extends React.Component {
       {
         maxZoom: 19,
         attribution: 'Carto © CC BY 3.0, OpenStreetMap © ODbL',
+        print: false,
       },
     );
 
-    const cartoLabels = L.tileLayer(
+    const cartoLabels = L.tileLayer.makeLabelsReadable(
       'https://cartodb-basemaps-{s}.global.ssl.fastly.net/dark_only_labels/{z}/{x}/{y}.png',
       {
         maxZoom: 18,
         attribution: 'Carto © CC BY 3.0, OpenStreetMap © ODbL',
+        print: true,
       },
     );
-    // creates layers with contrast and brightness filters applied
-    cartoLabels.createTile = function(coords, done) {
-      const tile = L.DomUtil.create('canvas', 'leaflet-tile');
-      tile.width = tile.height = this.options.tileSize;
 
-      const imageObj = new Image();
-      imageObj.crossOrigin = '';
-      imageObj.onload = function() {
-        const ctx = tile.getContext('2d');
-        ctx.filter = 'contrast(180%)';
-        ctx.filter = 'brightness(1.6)';
-        ctx.drawImage(imageObj, 0, 0);
-        done(null, tile); // Syntax is 'done(error, tile)'
-      };
-      imageObj.onerror = function(error) {
-        done(error, null);
-      };
-      imageObj.src = this.getTileUrl(coords);
-      return tile;
-    };
+    const satelliteStreets = L.mapboxGL({
+      attribution:
+        '<a href="https://www.maptiler.com/license/maps/" target="_blank">© MapTiler</a> <a href="https://www.openstreetmap.org/copyright" target="_blank">© OpenStreetMap contributors</a>',
+      accessToken: MAPBOX_TOKEN,
+      style: roadStyle,
+      print: false,
+      preserveDrawingBuffer: true,
+    });
+
+    const satelliteAdmin = L.mapboxGL({
+      attribution:
+        '<a href="https://www.maptiler.com/license/maps/" target="_blank">© MapTiler</a> <a href="https://www.openstreetmap.org/copyright" target="_blank">© OpenStreetMap contributors</a>',
+      accessToken: MAPBOX_TOKEN,
+      style: boundariesStyle,
+      print: false,
+      preserveDrawingBuffer: true,
+    });
+
     this.mainMap = L.map(mapId, {
       center: [lat, lng],
       zoom: zoom,
       minZoom: 3,
       layers: [cartoVoyagerNoLabels],
+      maxBounds: L.latLngBounds(mapMaxBounds),
+      maxBoundsViscosity: 0.9, //bounce back effect
+      attributionControl: false,
     });
     this.mainMap.createPane(COMPARE_LAYER_PANE);
     this.mainMap.createPane(ACTIVE_LAYER_PANE);
     this.mainMap.createPane(EDIT_LAYER_PANE);
     this.mainMap.createPane(LABEL_LAYER_PANE);
+    this.mainMap.createPane(ADMIN_LAYER_PANE);
+    this.mainMap.createPane(ROAD_LAYER_PANE);
     this.mainMap.getPane(MARKER_LAYER_PANE).style.zIndex = 63;
-    this.mainMap.getPane(OVERLAY_LAYER_PANE).style.zIndex = 62;
     this.mainMap.getPane(EDIT_LAYER_PANE).style.zIndex = 62;
     this.mainMap.getPane(LABEL_LAYER_PANE).style.zIndex = 61;
+    this.mainMap.getPane(ROAD_LAYER_PANE).style.zIndex = 60;
+    this.mainMap.getPane(ADMIN_LAYER_PANE).style.zIndex = 60;
     this.mainMap.getPane(COMPARE_LAYER_PANE).style.zIndex = 50;
     this.mainMap.getPane(ACTIVE_LAYER_PANE).style.zIndex = 40;
     this.mainMap.getPane(ACTIVE_LAYER_PANE).style.pointerEvents = 'none';
@@ -133,15 +158,21 @@ class RootMap extends React.Component {
 
     cartoLabels.options.pane = LABEL_LAYER_PANE;
     cartoLabels.addTo(this.mainMap);
+    satelliteStreets.options.pane = ROAD_LAYER_PANE;
+    satelliteAdmin.options.pane = ADMIN_LAYER_PANE;
 
     const baseMaps = {
       'Carto Voyager': cartoVoyagerNoLabels,
       'Carto Light': cartoLightNoLabels,
     };
     const overlays = {
+      Roads: satelliteStreets,
+      Borders: satelliteAdmin,
       Labels: cartoLabels,
     };
-
+    this.ruler = L.ruler().addTo(this.mainMap);
+    L.control.scale({ imperial: false, position: 'bottomright' }).addTo(this.mainMap);
+    L.control.attribution({ position: 'bottomleft' }).addTo(this.mainMap);
     this.poiLayer = L.geoJson().addTo(this.mainMap);
     this.poiLayer.pm.enable();
 
@@ -157,9 +188,15 @@ class RootMap extends React.Component {
         },
       })
       .addTo(this.mainMap);
-    L.control.scale({ imperial: false }).addTo(this.mainMap);
 
     this.mainMap.zoomControl.setPosition('bottomright');
+
+    this.mainMap.on('layeradd', addedLayer => {
+      if (addedLayer.layer instanceof L.MapboxGL) {
+        addedLayer.layer._update();
+      }
+    });
+
     this.mainMap.on('moveend', () => {
       Store.setMapView({
         lat: this.mainMap.getCenter().lat,
@@ -169,9 +206,10 @@ class RootMap extends React.Component {
       Store.setMapBounds(this.mainMap.getBounds(), this.mainMap.getPixelBounds());
     });
     this.mainMap.on('mousemove', e => {
+      const coords = e.latlng.wrap();
       const zoom = this.mainMap.getZoom();
-      const lng = roundDegrees(e.latlng.lng, zoom);
-      const lat = roundDegrees(e.latlng.lat, zoom);
+      const lng = roundDegrees(coords.lng, zoom);
+      const lat = roundDegrees(coords.lat, zoom);
       const value = `Lat: ${lat}, Lng: ${lng}`;
       document.getElementById('mapCoords').innerHTML = value;
     });
@@ -179,10 +217,27 @@ class RootMap extends React.Component {
     this.mainMap.on('resize', () => {
       this.resetAllLayers();
       if (Store.current.compareModeType === 'split') {
-        Store.current.pinResults.forEach((pin, index) => {
+        this.props.pins.forEach((pin, index) => {
           this.setOverlayParams(pin.opacity, index);
         });
       }
+    });
+
+    this.mainMap.on('measure:startMeasure', e => {
+      this.setState({ hasMeasurement: true, measureDistance: null, measureArea: null });
+    });
+    //measure events
+    this.mainMap.on('measure:move', e => {
+      this.setState({ measureDistance: e.distance, measureArea: e.area });
+    });
+    this.mainMap.on('measure:pointAdded', e => {
+      this.setState({ measureDistance: e.distance, measureArea: e.area });
+    });
+    this.mainMap.on('measure:finish', e => {
+      this.setState({ measureDistance: e.distance, measureArea: e.area });
+    });
+    this.mainMap.on('measure:removed', e => {
+      this.setState({ hasMeasurement: false, measureDistance: null, measureArea: null });
     });
 
     // add leaflet.pm controls to the map
@@ -228,12 +283,10 @@ class RootMap extends React.Component {
     this.aoiLayer.addTo(this.mainMap);
 
     this.mainMap.setView([Store.current.lat, Store.current.lng], Store.current.zoom);
+    this.mainMap.trackResize = true;
     Store.setMapBounds(this.mainMap.getBounds(), this.mainMap.getPixelBounds());
     this.setState({ isLoaded: true });
     this.visualizeLayer();
-    window.addEventListener('resize', () => {
-      this.invalidateMapSize({ pan: false });
-    });
   }
 
   toggleMapEdit = () => {
@@ -244,6 +297,14 @@ class RootMap extends React.Component {
           allowSelfIntersection: true,
         })
       : this.mainMap.pm.disableDraw('Poly');
+  };
+
+  toggleMeasure = () => {
+    this.ruler.toggle();
+  };
+
+  removeMeasurement = () => {
+    this.ruler.removeMeasurement();
   };
 
   removeAOILayers = () => {
@@ -331,10 +392,13 @@ class RootMap extends React.Component {
       this.removeAOILayers();
     }
     if (sCompareMode !== prevProps.compareMode) {
-      this.addPinLayers(Store.current.compareMode);
+      this.addPinLayers();
     }
     if (Store.current.compareMode && compareModeType !== Store.current.compareModeType) {
       this.compareLayers.forEach(this.resetLayerParams);
+    }
+    if (this.props.mapMaxBounds !== prevProps.mapMaxBounds) {
+      this.mainMap.setMaxBounds(L.latLngBounds(this.props.mapMaxBounds));
     }
   }
 
@@ -373,64 +437,51 @@ class RootMap extends React.Component {
 
   showPolygons(data) {
     this.clearPolygons();
-    Object.keys(Store.current.searchResults).map(
-      ds => (this.allResults = [...this.allResults, ...Store.current.searchResults[ds]]),
-    );
-    this.polyHighlight = L.geoJSON(this.allResults, {
-      style: Store.current.defaultPolyStyle,
-      onEachFeature: this.onEachPolygon,
+    this.allResults = [...Store.current.searchResults[DEFAULT_RESULTS_GROUP]];
+    // see https://leafletjs.com/examples/geojson/ for GeoJSON format
+    const geoJsonList = this.allResults.filter(r => r.tileData.dataGeometry).map((r, resultIndex) => ({
+      type: 'Feature',
+      properties: {
+        sensingTime: r.tileData.sensingTime,
+        resultIndex: r.resultIndex,
+      },
+      geometry: r.tileData.dataGeometry,
+    }));
+
+    this.polyHighlight = L.geoJSON(geoJsonList, {
+      style: DEFAULT_POLY_STYLE,
+      onEachFeature: (feature, layer) => {
+        layer.on({
+          mouseover: this.setHighlight,
+          mouseout: this.resetHighlight,
+          click: this.selectFeature,
+        });
+      },
     });
     this.polyHighlight.addTo(this.mainMap);
-    if (Store.current.searchParams.queryBounds !== undefined) {
-      this.boundPoly = L.polyline(getCoordsFromBounds(Store.current.searchParams.queryBounds, true), {
-        color: '#c1d82d',
-        dasharray: '5,5',
-        weight: 3,
-        fillColor: 'none',
-      }).addTo(this.mainMap);
-    }
   }
 
   resetHighlight = () => {
-    this.polyHighlight.setStyle(Store.current.defaultPolyStyle);
+    this.polyHighlight.setStyle(DEFAULT_POLY_STYLE);
   };
 
-  highlightFeature = e => {
-    this.resetHighlight(e);
-    let layer = e.target;
-    layer.setStyle(Store.current.highlightPolyStyle);
-    if (e.originalEvent.type === 'click') {
-      let obj = {};
-      let allResults = [];
-      Store.current.datasources.forEach(ds => (obj[ds] = [])); //we prepare object with all searched datasources
-      let p = [e.latlng.lng, e.latlng.lat];
-      this.allResults.forEach(layer => {
-        if (
-          gju.pointInPolygon(
-            {
-              type: 'Point',
-              coordinates: p,
-            },
-            layer.geometry,
-          )
-        ) {
-          obj[layer.tileData.datasource].push(layer);
-          allResults.push(layer);
-        }
-      });
-      if (allResults.length >= 1) {
-        layer.closePopup();
-        Store.setFilterResults(obj);
-      }
+  setHighlight = e => {
+    const layer = e.target;
+    layer.setStyle(HIGHLIGHT_POLY_STYLE);
+  };
+
+  selectFeature = e => {
+    const clickCoords = {
+      type: 'Point',
+      coordinates: [e.latlng.lng, e.latlng.lat],
+    };
+    const intersectingResults = this.allResults.filter(
+      r => r.tileData.dataGeometry && gju.pointInPolygon(clickCoords, r.tileData.dataGeometry),
+    );
+
+    if (intersectingResults.length > 0) {
+      Store.setFilterResults({ DEFAULT_RESULTS_GROUP: intersectingResults });
     }
-  };
-
-  onEachPolygon = (feature, layer) => {
-    layer.on({
-      mouseover: this.highlightFeature,
-      mouseout: this.resetHighlight,
-      click: this.highlightFeature,
-    });
   };
 
   clearPolygons = () => {
@@ -477,16 +528,25 @@ class RootMap extends React.Component {
     this.mainMap.setView([lat, lng]);
   };
 
-  highlightTile = index => {
-    if (this.polyHighlight) this.resetHighlight();
-
-    let layersArray = [];
-    if (this.polyHighlight && this.polyHighlight._layers !== undefined)
-      layersArray = Object.keys(this.polyHighlight._layers);
-
-    if (this.polyHighlight && this.polyHighlight._layers !== undefined) {
-      this.polyHighlight._layers[layersArray[index]].setStyle(Store.current.highlightPolyStyle);
+  highlightTile = resultIndex => {
+    // There is sometimes a problem with highlights (there aren't any)
+    // when Object.values(this.polyHighlight._layers).length is 0 (zero).
+    // Some of our datasources apparently don't have this feature (Envisat Meris),
+    // or there is a problem with getting data from backend.
+    // So checking [if length of (object converted to array) is zero] is added.
+    if (
+      !this.polyHighlight ||
+      !this.polyHighlight._layers ||
+      Object.values(this.polyHighlight._layers).length === 0
+    ) {
+      return;
     }
+    this.resetHighlight();
+
+    const layerFeature = Object.values(this.polyHighlight._layers).find(
+      l => l.feature.properties.resultIndex === resultIndex,
+    );
+    layerFeature.setStyle(HIGHLIGHT_POLY_STYLE);
   };
 
   onZoomToPin(item) {
@@ -530,9 +590,12 @@ class RootMap extends React.Component {
     this.compareLayers = [];
   };
   drawCompareLayers = () => {
-    let pins = [...Store.current.pinResults];
+    let pins = [...this.props.pins];
+    let { instances, compareMode: isCompare } = Store.current;
+    if (!isCompare) {
+      return;
+    }
     pins.reverse().forEach(item => {
-      let { instances } = Store.current;
       let layer = createMapLayer(
         instances.find(inst => inst.name === item.datasource),
         COMPARE_LAYER_PANE,
@@ -562,10 +625,6 @@ class RootMap extends React.Component {
     this.drawCompareLayers();
   };
 
-  invalidateMapSize = options => {
-    this.mainMap.invalidateSize(options);
-  };
-  // App.js handles in onCompareChange
   addPinLayers = () => {
     const { mainTabIndex: tabIndex, compareMode: isCompare } = Store.current;
 
@@ -573,7 +632,7 @@ class RootMap extends React.Component {
       if (this.activeLayer !== null) {
         this.mainMap.removeLayer(this.activeLayer);
       }
-      let pins = [...Store.current.pinResults];
+      let pins = [...this.props.pins];
       pins.reverse().forEach(item => {
         let { instances } = Store.current;
         let layer = createMapLayer(
@@ -602,7 +661,7 @@ class RootMap extends React.Component {
   setOverlayParams = (arr, index) => {
     //if not in compare mode, don't do anything
     if (!Store.current.compareMode) return;
-    let mapIndex = Store.current.pinResults.length - (index + 1);
+    let mapIndex = this.props.pins.length - (index + 1);
     if (Store.current.compareModeType === 'opacity') {
       this.compareLayers[mapIndex].setOpacity(arr[1]);
     } else {
@@ -620,30 +679,79 @@ class RootMap extends React.Component {
 
   getUpdateParams(item) {
     let { selectedResult, presets, compareMode, isEvalUrl } = Store.current;
-    let { datasource, gain, gamma, time, evalscript, evalscripturl, atmFilter, preset } =
+    let {
+      datasource: datasourceName,
+      gainOverride,
+      gammaOverride,
+      redRangeOverride,
+      greenRangeOverride,
+      blueRangeOverride,
+      valueRangeOverride,
+      time,
+      evalscript,
+      evalscripturl,
+      atmFilter,
+      preset: layerId,
+    } =
       item || selectedResult || {};
-    if (!datasource) return;
+
+    // datasource is just a string (name of selected datasource/satellite)
+    // we need data of the selected datasource from the presets in Store
+    let datasourceLayers = presets[datasourceName];
+
+    // preset is just a string (name of selected layer)
+    // item is usually data of the selected layer from config.js (but apparently not always - not when Pins made from Results are being compared)
+    // selectedResult is just data of the selected layer from config.js
+    // item and selectedResult are basically the same if it's not a custom layer
+    // we need data of the selected layer from the presets in Store
+    let selectedLayer = datasourceLayers.find(l => l.id === layerId);
+
+    if (!datasourceName) return;
     let obj = {};
-    obj.format = 'image/png';
+    obj.format = selectedLayer && selectedLayer.format ? selectedLayer.format : 'image/png';
     obj.pane = compareMode ? COMPARE_LAYER_PANE : ACTIVE_LAYER_PANE;
     obj.transparent = true;
     obj.maxcc = 100;
-    if (datasource.includes('EW') && preset.includes('NON_ORTHO')) {
+    if (datasourceName.includes('EW') && layerId.includes('NON_ORTHO')) {
       obj.orthorectify = false;
     }
-    if (isCustomPreset(preset)) {
-      obj.layers = presets[datasource][0].id;
+
+    if (isCustomPreset(layerId)) {
+      selectedResult && selectedResult.baseUrls.WMTS
+        ? (obj.layer = presets[datasourceName][0].id)
+        : (obj.layers = presets[datasourceName][0].id);
+
       evalscripturl && isEvalUrl && (obj.evalscripturl = evalscripturl);
       !isEvalUrl && (obj.evalscript = evalscript);
-      obj.evalsource = evalSourcesMap[datasource];
+      obj.evalsource = evalSourcesMap[datasourceName];
       obj.PREVIEW = 3;
     } else {
-      obj.layers = preset;
+      selectedResult && selectedResult.baseUrls.WMTS ? (obj.layer = layerId) : (obj.layers = layerId);
     }
-    gain && (obj.gain = gain);
-    gamma && (obj.gamma = gamma);
+
+    const evalscriptoverridesObj = {
+      gainOverride,
+      gammaOverride,
+      redRangeOverride,
+      greenRangeOverride,
+      blueRangeOverride,
+      valueRangeOverride,
+    };
+    obj.evalscriptoverrides = btoa(evalscriptoverridesToString(evalscriptoverridesObj));
+
     atmFilter && (obj.ATMFILTER = atmFilter === 'null' ? null : atmFilter);
-    obj.time = `${time}/${time}`;
+
+    if (item.baseUrls && item.baseUrls.WMTS) {
+      obj.time = `${time}`;
+      obj.tilematrixSet = selectedLayer.tilematrixSetName;
+      obj.minZoom = selectedLayer.minZoom;
+      obj.maxZoom = selectedLayer.maxZoom;
+    } else {
+      obj.time = `${time}/${time}`;
+    }
+    if (selectedResult && typeof selectedResult.defaultStyle === 'function') {
+      obj.styles = selectedResult.defaultStyle(layerId);
+    }
     return cloneDeep(obj);
   }
 
@@ -658,7 +766,7 @@ class RootMap extends React.Component {
     this.mainMap.fitBounds(this.aoiLayer.getBounds());
     this.setState({ uploadDialog: false });
   };
-  openUploadKMLDialog = () => {
+  openUploadGeoFileDialog = () => {
     this.setState({ uploadDialog: true });
   };
   drawMarker = () => {
@@ -685,26 +793,36 @@ class RootMap extends React.Component {
     this.setState({ fisDialog: true, fisDialogAoiOrPoi: aoiOrPoi });
   };
   render() {
-    const { uploadDialog, fisDialog, fisDialogAoiOrPoi } = this.state;
-    const { aoiBounds, poi, mapGeometry, isAoiClip, user, selectedResult } = Store.current;
+    const {
+      uploadDialog,
+      fisDialog,
+      fisDialogAoiOrPoi,
+      hasMeasurement,
+      measureDistance,
+      measureArea,
+    } = this.state;
+    const { aoiBounds, poi, mapGeometry, isAoiClip, user, selectedResult } = this.props;
     return (
-      <div className="map-wrapper" style={{ width: this.props.width }}>
+      <div className="map-wrapper">
         <div className="mainMap" id={this.props.mapId} />
         {uploadDialog && (
-          <UploadKML onUpload={this.onUpload} onClose={() => this.setState({ uploadDialog: false })} />
+          <UploadGeoFile onUpload={this.onUpload} onClose={() => this.setState({ uploadDialog: false })} />
         )}
         <div id="aboutSentinel">
           <a href="https://www.sentinel-hub.com/apps/eo_browser" target="_blank" rel="noopener noreferrer">
             About EO Browser
           </a>
-          <a href="mailto:info@sentinel-hub.com?Subject=EO%20Browser%20Feedback">Contact us</a>
+          <a href="https://forum.sentinel-hub.com/" target="_blank" rel="noopener noreferrer">
+            Contact us
+          </a>
           <a href="https://www.sentinel-hub.com/apps/wms" target="_blank" rel="noopener noreferrer">
             Get data
           </a>
         </div>
         <div id="mapCoords" />
         <div>
-          <DrawArea
+          <InfoPanelButton />
+          <AOIPanelButton
             fisDialog={fisDialog}
             aoiBounds={aoiBounds}
             isAoiClip={isAoiClip}
@@ -712,11 +830,11 @@ class RootMap extends React.Component {
             selectedResult={selectedResult}
             openFisPopup={this.openFisPopup}
             resetAoi={this.resetAoi}
-            openUploadKMLDialog={this.openUploadKMLDialog}
+            openUploadGeoFileDialog={this.openUploadGeoFileDialog}
             centerOnFeature={this.centerOnFeature}
             disabled={!user}
           />
-          <DrawMarker
+          <POIPanelButton
             openFisPopup={this.openFisPopup}
             drawMarker={this.drawMarker}
             poi={poi}
@@ -726,6 +844,8 @@ class RootMap extends React.Component {
             centerOnFeature={this.centerOnFeature}
             disabled={!user}
           />
+          <DownloadPanelButton originalMap={this.mainMap} />
+          <TimelapsePanelButton />
 
           {fisDialog && (
             <FIS
@@ -734,10 +854,39 @@ class RootMap extends React.Component {
               onClose={() => this.setState({ fisDialog: false })}
             />
           )}
+          <MeasurePanelButton
+            toggleMeasure={this.toggleMeasure}
+            hasMeasurement={hasMeasurement}
+            distance={measureDistance}
+            area={measureArea}
+            removeMeasurement={this.removeMeasurement}
+            isLoggedIn={user}
+          />
         </div>
       </div>
     );
   }
 }
 
-export default connect(store => store, null, null, { withRef: true })(RootMap);
+const mapStoreToProps = store => ({
+  action: store.action,
+  aoiBounds: store.aoiBounds,
+  compareMode: store.compareMode,
+  compareModeType: store.compareModeType,
+  formQuery: store.formQuery,
+  instances: store.instances,
+  isAoiClip: store.isAoiClip,
+  lat: store.lat,
+  lng: store.lng,
+  location: store.location,
+  mainTabIndex: store.mainTabIndex,
+  mapGeometry: store.mapGeometry,
+  pins: store.themePins || store.userPins,
+  poi: store.poi,
+  selectedResult: store.selectedResult,
+  updatePosition: store.updatePosition,
+  user: store.user,
+  zoom: store.zoom,
+  mapMaxBounds: store.mapMaxBounds,
+});
+export default connect(mapStoreToProps, null, null, { withRef: true })(RootMap);
