@@ -2,28 +2,83 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import './Results.scss';
 import ResultsInnerPanel from './ResultsInnerPanel';
-import get from 'dlv';
+import Store from '../../store';
+import { performSearch, currentSearchHasMoreResults } from '../../datasources';
+import { DEFAULT_RESULTS_GROUP } from '../RootMap';
+import { NotificationPanel } from 'eo-components';
 
 class ResultsPanel extends React.Component {
-  renderDatasourceResults = (ds, i) => {
-    let items = get(this.props.results, ds) || [];
+  constructor(props) {
+    super(props);
+    this.state = {
+      searchError: null,
+      searchInProgress: false,
+    };
+  }
+
+  loadMore = () => {
+    this.setState({
+      searchError: null,
+      searchInProgress: true,
+    });
+
+    performSearch(Store.current.lat, Store.current.lng)
+      .then(finalResults => {
+        const mergedResults = [...Store.current.searchResults[DEFAULT_RESULTS_GROUP], ...finalResults];
+        Store.setSearchResults(mergedResults, DEFAULT_RESULTS_GROUP, {});
+        this.props.onFinishSearch(Store.current.searchResults);
+      })
+      .catch(errMsg => {
+        this.setState({
+          searchError: errMsg,
+        });
+        this.props.onFinishSearch();
+      })
+      .then(() => {
+        this.setState({
+          searchInProgress: false,
+        });
+      })
+      .catch(e => {
+        console.error(e);
+      });
+  };
+
+  renderDatasourceResults = () => {
+    // Why is this.props.results an object with datasource names as keys? Good question. One
+    // might also ask why there are datasource and datasources (both, and at all) in props... Anyway. :)
+    const {
+      results,
+      searchParams,
+      showClear,
+      onResultClick,
+      onResultHover,
+      allowLoadMoreButton,
+    } = this.props;
+    const { searchInProgress } = this.state;
+    if (!results || Object.keys(results).length === 0) {
+      return null;
+    }
+
+    let items = Object.values(results)[0];
     if (items.length === 0) {
       return null;
     }
     return (
-      items.length > 0 && (
+      <div>
         <ResultsInnerPanel
-          key={i}
-          isSearching={this.props.isSearching}
-          datasource={ds}
-          searchParams={this.props.searchParams}
-          showClear={this.props.showClear}
-          onResultClick={this.props.onResultClick}
-          onResultHover={this.props.onResultHover}
-          onLoadMore={this.props.onLoadMore}
+          searchParams={searchParams}
+          showClear={showClear}
+          onResultClick={onResultClick}
+          onResultHover={onResultHover}
+          onLoadMore={this.loadMore}
+          showLoadMoreButton={
+            allowLoadMoreButton && !this.state.searchInProgress && currentSearchHasMoreResults()
+          }
           results={items}
         />
-      )
+        {searchInProgress && <NotificationPanel msg="Loading more results ..." type="loading" />}
+      </div>
     );
   };
 
@@ -36,8 +91,7 @@ class ResultsPanel extends React.Component {
             Clear data
           </a>
         )}
-        {this.props.datasources &&
-          this.props.datasources.map((inst, i) => this.renderDatasourceResults(inst, i))}
+        {this.renderDatasourceResults(this.props.results)}
       </div>
     );
   }
@@ -51,9 +105,8 @@ ResultsPanel.propTypes = {
   onResultHover: PropTypes.func,
   searchParams: PropTypes.object,
   showClear: PropTypes.bool,
-  isSearching: PropTypes.bool,
-  onLoadMore: PropTypes.func,
   onClearData: PropTypes.func,
+  onFinishSearch: PropTypes.func,
 };
 
 export default ResultsPanel;
