@@ -1,51 +1,41 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import './Results.scss';
-import NotificationPanel from '../NotificationPanel';
 import ResultItem from './ResultItem';
-import get from 'dlv';
 
 const style = { clear: 'both' };
 class ResultsInnerPanel extends React.Component {
   renderItems = results => {
     return (
       <div style={style}>
-        {results.map((result, i) => {
-          return (
-            <ResultItem
-              key={i}
-              result={result}
-              onResultClick={this.props.onResultClick}
-              onResultHover={this.props.onResultHover}
-            />
-          );
-        })}
+        {results.map((result, i) => (
+          <ResultItem
+            key={`${result.tileData.datasource}.${result.tileData.id ? result.tileData.id : i}`}
+            result={result}
+            onResultClick={this.props.onResultClick}
+            onResultHover={this.props.onResultHover}
+          />
+        ))}
       </div>
     );
   };
 
-  renderWaypoint = (ds, isBig) => {
-    if (!this.props.isSearching && get(this.props, `searchParams.${ds}.hasMore`)) {
+  renderLoadMoreButton = () => {
+    // we only display first 50 results, but the 51st result tells us that there would be more available if we wanted them:
+    if (this.props.showLoadMoreButton) {
       return (
-        <a className={isBig && 'btn'} onClick={() => this.props.onLoadMore(ds)}>
+        <a className="btn" onClick={() => this.props.onLoadMore()}>
           Load more
         </a>
       );
     }
   };
 
-  renderHeading = (results, datasource) => {
-    let { datasource: name } = results[0].tileData; //we read name from first item
+  renderHeading = results => {
     return (
       <div>
         <div className="resultsHeading">
-          {name || datasource}: Showing {results.length} {results.length > 1 ? 'results' : 'result'}.
-          {this.renderWaypoint(datasource, false)}
-          {this.props.isSearching && (
-            <a>
-              <i className="fa fa-spinner fa-spin fa-fw" />
-            </a>
-          )}
+          Showing {results.length} {results.length === 1 ? 'result' : 'results'}.
           <div style={{ clear: 'both' }} />
         </div>
       </div>
@@ -53,14 +43,16 @@ class ResultsInnerPanel extends React.Component {
   };
 
   render() {
-    let { results, datasource } = this.props;
-    if (results.length === 0) return null;
+    let { results } = this.props;
+    if (results.length === 0) {
+      return null;
+    }
     return (
       <div className="resultsInnerPanel">
-        {datasource && this.renderHeading(results, datasource)}
+        {this.renderHeading(results)}
         {this.renderItems(results)}
-        {this.renderWaypoint(datasource, true)}
-        {this.props.isSearching && <NotificationPanel msg="Loading more results ..." type="loading" />}
+        {this.renderLoadMoreButton()}
+        {/* {isSearching && <NotificationPanel msg="Loading more results ..." type="loading" />} */}
       </div>
     );
   }
@@ -72,9 +64,8 @@ ResultsInnerPanel.propTypes = {
   onResultHover: PropTypes.func,
   searchParams: PropTypes.object,
   showClear: PropTypes.bool,
-  datasource: PropTypes.string,
-  isSearching: PropTypes.bool,
   onLoadMore: PropTypes.func,
+  showLoadMoreButton: PropTypes.bool,
 };
 
 export default ResultsInnerPanel;
