@@ -4,9 +4,9 @@ import Store from '../store';
 import { getCoordsFromBounds } from './coords';
 
 export const ISO_8601_UTC = `YYYY-MM-DD[T]HH:mm:ss.SSS[Z]`;
-
+export const STANDARD_STRING_DATE_FORMAT = 'YYYY-MM-DD';
 export function getAndSetNextPrev(direction) {
-  const { maxDate, minDate, selectedResult, dateFormat } = Store.current;
+  const { maxDate, minDate, selectedResult } = Store.current;
   let { time, datasource, indexService } = selectedResult;
   let { mapBounds: bounds } = Store.current;
   let activeLayer = [...Store.current.instances, ...(Store.current.userInstances || {})].find(inst => {
@@ -20,11 +20,14 @@ export function getAndSetNextPrev(direction) {
   };
   const from =
     direction === 'prev'
-      ? moment(minDate).format(dateFormat)
+      ? moment(minDate).format(STANDARD_STRING_DATE_FORMAT)
       : moment(time)
           .add(1, 'day')
-          .format(dateFormat);
-  const to = direction === 'prev' ? moment(time).format(dateFormat) : moment(maxDate).format(dateFormat);
+          .format(STANDARD_STRING_DATE_FORMAT);
+  const to =
+    direction === 'prev'
+      ? moment(time).format(STANDARD_STRING_DATE_FORMAT)
+      : moment(maxDate).format(STANDARD_STRING_DATE_FORMAT);
   let payload = {
     clipping: clip,
     maxcount: 1,
@@ -58,21 +61,15 @@ export function getAndSetNextPrev(direction) {
 }
 
 export function queryDatesForActiveMonth(singleDate, datasourceName) {
-  const startOfMonth = moment(singleDate)
-    .startOf('month')
-    .format(ISO_8601_UTC);
-  const endOfMonth = moment(singleDate)
-    .endOf('month')
-    .format(ISO_8601_UTC);
+  const startOfMonth = moment(singleDate).startOf('month');
+  const endOfMonth = moment(singleDate).endOf('month');
   return fetchAvailableDates(startOfMonth, endOfMonth, datasourceName);
 }
 
 export function fetchAvailableDates(from, to, selectedDataSource, queryArea) {
   let selectedResult =
     Store.current.selectedResult || Store.current.instances.find(ds => ds.name === selectedDataSource);
-
   const newService = selectedResult.indexService && selectedResult.indexService.includes('v3/collections');
-
   if (!queryArea) {
     queryArea = {
       type: 'Polygon',
@@ -85,47 +82,6 @@ export function fetchAvailableDates(from, to, selectedDataSource, queryArea) {
       coordinates: [getCoordsFromBounds(Store.current.mapBounds, false, newService)],
     };
   }
-
-  if (selectedResult.getDates) {
-    return selectedResult.getDates(from, to, queryArea);
-  }
-
-  let url;
-  let payload;
-  if (newService) {
-    payload = {
-      queryArea: queryArea,
-      from: from,
-      to: to,
-      maxCloudCoverage: 1,
-    };
-    url = selectedResult.dateService;
-  } else {
-    payload = queryArea;
-    url = `${selectedResult.dateService}?timefrom=${from}&timeto=${to}'`;
-  }
-
-  return fetchDatesFromDatesService(url, payload);
-}
-
-function fetchDatesFromDatesService(url, payload) {
-  return new Promise((resolve, reject) => {
-    axios
-      .post(url, payload, {
-        headers: new Headers({
-          'Accept-CRS': 'EPSG:4326',
-          'Content-Type': 'application/json;charset=utf-8',
-          Accept: 'application/json',
-        }),
-      })
-      .then(response => response.data)
-      .then(res => {
-        if (res.length > 0) {
-          resolve(res);
-        }
-        if (res.length === 0) {
-          reject('no date found');
-        }
-      });
-  });
+  const getDatesMethod = selectedResult.getDates || selectedResult.activeLayer.getDates;
+  return getDatesMethod(from, to, queryArea);
 }
