@@ -1,8 +1,14 @@
+import axios from 'axios';
 import values from 'lodash/values';
+import cloneDeep from 'lodash/cloneDeep';
 import isEqual from 'lodash/isEqual';
 import L from 'leaflet';
 import '../components/ext/leaflet-clip-wms-layer';
 import Store from '../store';
+import { getTokenFromLocalStorage, isTokenExpired } from './auth';
+import { getFisShadowLayer } from '../components/FIS';
+
+import '../components/ext/leaflet-tilelayer-wmts-src';
 
 export function getMultipliedLayers(layers) {
   let result = [];
@@ -26,23 +32,6 @@ export function getCrsLabel(tile) {
   return { crs, mgrs };
 }
 
-export const evalSourcesMap = {
-  'Sentinel-1': 'S1',
-  'Sentinel-1 GRD IW': 'S1',
-  'Sentinel-1 GRD EW': 'S1_EW',
-  'Sentinel-1 GRD EW SH': 'S1_EW_SH',
-  'Sentinel-2 L1C': 'S2',
-  MODIS: 'Modis',
-  'Sentinel-2': 'S2',
-  'Sentinel-2 L2A': 'S2L2A',
-  'Sentinel-3 OLCI': 'S3',
-  'Landsat 5 ESA': 'L5',
-  'Landsat 7 ESA': 'L7',
-  'Landsat 8': 'L8',
-  'Landsat 8 ESA': 'L8',
-  'Landsat 8 USGS': 'L8',
-  'Envisat Meris': 'ENV',
-};
 export function isCustomPreset(preset) {
   return preset === 'CUSTOM';
 }
@@ -52,15 +41,87 @@ export function getLayersString(layers) {
 }
 
 export function hasPinSaved(currPin) {
-  const { pinResults } = Store.current;
-  return pinResults.find(pin => {
+  const { userPins } = Store.current;
+  return userPins.find(pin => {
     return isEqual(pin, currPin);
   });
 }
 
+export function getZoomLimitsForSelectedLayer() {
+  const defaultMinZoom = 5;
+  const defaultMaxZoom = 16;
+  //const defaultAllowOverZoomBy = 0
+
+  let minZoomFromGetCapabilities = undefined;
+  let maxZoomFromGetCapabilities = undefined;
+
+  if (Store.current.selectedResult) {
+    const selectedDatasourceLayers = Store.current.presets[Store.current.selectedResult.name];
+    const selectedLayer = selectedDatasourceLayers
+      ? selectedDatasourceLayers.find(l => l.id === Store.current.selectedResult.preset)
+      : undefined;
+    if (selectedLayer) {
+      minZoomFromGetCapabilities = selectedLayer.minZoom;
+      maxZoomFromGetCapabilities = selectedLayer.maxZoom;
+    }
+  }
+  // no allowOverZoomBy (or similar) from getCapabilities
+
+  let minZoomFromConfig = undefined;
+  let maxZoomFromConfig = undefined;
+  //let allowOverZoomByFromConfig = undefined;
+  if (Store.current.selectedResult) {
+    const selectedDatasourceFromConfig = Store.current.instances.find(
+      d => d.name === Store.current.selectedResult.name,
+    );
+    minZoomFromConfig = selectedDatasourceFromConfig.minZoom;
+    maxZoomFromConfig = selectedDatasourceFromConfig.maxZoom;
+    //allowOverZoomByFromConfig = selectedDatasourceFromConfig.allowOverZoomBy;
+  }
+
+  const minZoom =
+    minZoomFromGetCapabilities !== undefined
+      ? minZoomFromGetCapabilities
+      : minZoomFromConfig !== undefined
+        ? minZoomFromConfig
+        : defaultMinZoom;
+  const maxZoom =
+    maxZoomFromGetCapabilities !== undefined
+      ? maxZoomFromGetCapabilities
+      : maxZoomFromConfig !== undefined
+        ? maxZoomFromConfig
+        : defaultMaxZoom;
+
+  // const allowOverZoomBy =
+  //   allowOverZoomByFromConfig !== undefined ? allowOverZoomByFromConfig : defaultAllowOverZoomBy;
+  const allowOverZoomBy = 2;
+
+  return { minZoom, maxZoom, allowOverZoomBy };
+}
+
 export function createMapLayer(instanceObj, pane, progress) {
   if (instanceObj === undefined) return;
-  const { name, baseUrls, minZoom = 5, maxZoom = 16, tileSize = 512 } = instanceObj;
+
+  let selectedLayerTileSize = 512;
+
+  if (Store.current.selectedResult) {
+    const selectedDatasourceLayers = Store.current.presets[instanceObj.name];
+    let selectedLayer = selectedDatasourceLayers
+      ? selectedDatasourceLayers.find(l => l.id === Store.current.selectedResult.preset)
+      : undefined;
+    selectedLayerTileSize = selectedLayer && selectedLayer.tileSize ? selectedLayer.tileSize : 512;
+  }
+
+  const {
+    name,
+    baseUrls,
+    tileSize = selectedLayerTileSize,
+    tilematrixSet = 'PopularWebMercator512',
+  } = instanceObj;
+
+  const { minZoom, maxZoom: maxNativeZoom, allowOverZoomBy } = getZoomLimitsForSelectedLayer();
+  // const maxNativeZoom = maxZoom;
+
   // when we create compare layer, we will create CLIP layer, otherwise normal wms layer
   let layer =
     pane === 'compareLayer'
@@ -68,96 +129,248 @@ export function createMapLayer(instanceObj, pane, progress) {
           showlogo: false,
           tileSize,
           minZoom,
-          maxZoom,
+          maxZoom: maxNativeZoom + allowOverZoomBy,
+          maxNativeZoom: maxNativeZoom,
           pane,
           name,
         })
-      : L.tileLayer.wms(baseUrls.WMS, {
-          showlogo: false,
-          tileSize,
-          minZoom,
-          maxZoom,
-          pane,
-          name,
-        });
+      : baseUrls.WMTS
+        ? L.tileLayer.wmts(baseUrls.WMTS, {
+            showlogo: false,
+            style: 'default',
+            tileSize,
+            minZoom,
+            maxZoom: maxNativeZoom + allowOverZoomBy,
+            maxNativeZoom: maxNativeZoom,
+            pane,
+            name,
+            tilematrixSet,
+          })
+        : L.tileLayer.wms(baseUrls.WMS, {
+            showlogo: false,
+            tileSize,
+            maxZoom: maxNativeZoom + allowOverZoomBy,
+            maxNativeZoom: maxNativeZoom,
+            minZoom,
+            pane,
+            name,
+          });
+
   layer.on('loading', function() {
     progress.start();
     progress.inc(0.1);
   });
+
   layer.on('load', function() {
     progress.done();
   });
-  return layer;
-}
 
-// Store pins to localStorage
-function hasLocalStorage() {
-  // inspired by Modernizr's test
-  // https://github.com/Modernizr/Modernizr/blob/5eea7e2a213edc9e83a47b6414d0250468d83471/feature-detects/storage/localstorage.js
-  // as of 02.03.2017
-  var mod = 'modernizr';
-  try {
-    localStorage.setItem(mod, mod);
-    localStorage.removeItem(mod);
-    return true;
-  } catch (e) {
-    return false;
-  }
+  layer.createTile = function(coords, done) {
+    const tile = document.createElement('img');
+    tile.width = tile.height = this.options.tileSize;
+    tile.setAttribute('data', 'retries');
+    tile.dataset.retries = 0;
+    tile.onload = function() {
+      done(null, tile); // Syntax is 'done(error, tile)'
+    };
+
+    tile.onerror = function(error) {
+      const src = tile.src;
+      setTimeout(() => {
+        if (tile.dataset.retries < 5) {
+          // retry img download if retries below 5
+          tile.src = '';
+          tile.dataset.retries++;
+          tile.src = src;
+        }
+      }, 1000);
+
+      done(error, tile); // Syntax is 'done(error, tile)'
+    };
+
+    tile.src = `${this.getTileUrl(coords)}`;
+    return tile;
+  };
+
+  return layer;
 }
 
 const PINS_LC_NAME = 'react-app-pins';
 
-function isValidId(id) {
+function isValidPinId(id) {
   const idRegex = /^[0-9].*-pin$/i;
   return idRegex.test(id);
 }
+
 export function uniquePinId() {
-  return `${Math.floor(Math.random() * 1000 + 100)}-${new Date().valueOf()}-pin`;
+  const token = getTokenFromLocalStorage();
+  const userPart = Store.current.user.sub
+    .split('-')
+    .reverse()
+    .join('')
+    .substring(0, 16);
+  const tokenPart = token.id_token.substring(token.id_token.length - 16, token.id_token.length);
+  return `${Math.floor(Math.random() * 1000 + 100)}-${userPart + tokenPart}-${new Date().valueOf()}-pin`;
+}
+
+export function fetchPinsFromServer() {
+  getPinsFromServer()
+    .then(serverPins => {
+      let shouldUpdateOnServer = false;
+      // at the same time, merge local pins: (if there are any)
+      const localPins = getPinsFromLocalStorage();
+      if (localPins.length === 0) {
+        return [serverPins, shouldUpdateOnServer];
+      }
+
+      shouldUpdateOnServer = true;
+      let mergedPins = [...serverPins];
+      localPins.forEach(lPin => {
+        if (!serverPins.find(sPin => sPin._id === lPin._id)) {
+          mergedPins.push(lPin);
+        }
+      });
+      return [mergedPins, shouldUpdateOnServer];
+    })
+    .then(([mergedPins, shouldUpdateOnServer]) => {
+      Store.setPins(mergedPins);
+      if (shouldUpdateOnServer) {
+        updatePinsOnServer(mergedPins)
+          .then(() => {
+            removePinsFromLocalStorage();
+          })
+          .catch(e => {
+            console.error(e);
+          });
+      }
+    })
+    .catch(e => {
+      if (e) {
+        console.error(e);
+      }
+    });
+}
+
+function getPinsFromServer() {
+  return new Promise((resolve, reject) => {
+    const token = getTokenFromLocalStorage();
+    if (!token || isTokenExpired(token)) {
+      reject();
+      return;
+    }
+    const url = `${process.env.REACT_APP_BASEURL}userdata/`;
+    const requestParams = {
+      responseType: 'json',
+      headers: {
+        Authorization: `Bearer ${token.access_token}`,
+      },
+    };
+    axios
+      .get(url, requestParams)
+      .then(res => {
+        try {
+          const serverPins = res.data.pins;
+          resolve(serverPins ? serverPins : []);
+        } catch (e) {
+          resolve([]);
+        }
+      })
+      .catch(e => {
+        if (e && e.response && e.response.status === 404) {
+          resolve([]);
+          return;
+        }
+        reject(e);
+      });
+  });
+}
+
+export async function updatePinsOnServer(pins) {
+  return new Promise((resolve, reject) => {
+    // get the original userData, fix it and upload again:
+    const token = getTokenFromLocalStorage();
+    if (!token) {
+      console.log('Unable to save pins, auth token not available');
+      reject();
+      return;
+    }
+    const url = `${process.env.REACT_APP_BASEURL}userdata/`;
+    const requestParams = {
+      responseType: 'json',
+      headers: {
+        Authorization: `Bearer ${token.access_token}`,
+      },
+    };
+    axios
+      .get(url, requestParams)
+      .then(res => res.data)
+      .catch(e => {
+        // if no user data is found, ignore the error:
+        if (e && e.response && e.response.status === 404) {
+          return {};
+        }
+        throw e;
+      })
+      .then(userData => {
+        const newData = cloneDeep(userData);
+        newData.pins = pins;
+        axios
+          .put(url, newData, requestParams)
+          .then(() => {
+            resolve(pins);
+          })
+          .catch(e => {
+            console.error('Unable to save pins!', e);
+            reject(e);
+          });
+      })
+      .catch(e => {
+        console.error('Unable to retrieve user data!', e);
+        reject(e);
+      });
+  });
 }
 
 export function getPinsFromLocalStorage() {
-  //return [];
-  if (!hasLocalStorage) return [];
-  let pins = window.localStorage.getItem(PINS_LC_NAME);
-  if (pins) {
-    return JSON.parse(pins)
-      .filter(pin => pin.name || pin.datasource)
-      .map((pin, index) => {
-        if (pin.properties) {
-          // legacy format
-          const {
-            name,
-            properties: {
-              rawData: { time },
-            },
-            map: { latitude, longitude, zoom },
-            ...rest
-          } = pin;
-          return {
-            ...rest,
-            _id: isValidId(pin._id) ? pin._id : uniquePinId(),
-            datasource: name,
-            lat: latitude,
-            lng: longitude,
-            zoom,
-            time,
-          };
-        }
-        return {
-          ...pin,
-          opacity: [0, 1],
-          _id: isValidId(pin._id) ? pin._id : uniquePinId(),
-        };
-      });
+  let pinsString = window.localStorage.getItem(PINS_LC_NAME);
+  if (!pinsString) {
+    return [];
   }
 
-  return [];
+  const localPins = JSON.parse(pinsString)
+    .filter(pin => pin.name || pin.datasource)
+    .map((pin, index) => {
+      if (pin.properties) {
+        // legacy format
+        const {
+          name,
+          properties: {
+            rawData: { time },
+          },
+          map: { latitude, longitude, zoom },
+          ...rest
+        } = pin;
+        return {
+          ...rest,
+          _id: isValidPinId(pin._id) ? pin._id : uniquePinId(),
+          datasource: name,
+          lat: latitude,
+          lng: longitude,
+          zoom,
+          time,
+        };
+      }
+      return {
+        ...pin,
+        opacity: [0, 1],
+        _id: isValidPinId(pin._id) ? pin._id : uniquePinId(),
+      };
+    });
+  return localPins;
 }
 
-export function syncPinsWithLocalStorage(pins) {
-  if (!hasLocalStorage) return;
-
-  window.localStorage.setItem(PINS_LC_NAME, JSON.stringify(pins));
+export function removePinsFromLocalStorage() {
+  window.localStorage.removeItem(PINS_LC_NAME);
 }
 
 export function b64DecodeUnicode(str) {
@@ -172,144 +385,59 @@ export function b64EncodeUnicode(str) {
   return btoa(str);
 }
 
-export function getPolyfill() {
-  if (!Array.prototype.includes) {
-    // prettier-ignore
-    Object.defineProperty(Array.prototype, 'includes', { // eslint-disable-line no-extend-native
-      value: function(searchElement, fromIndex) {
-        // 1. Let O be ? ToObject(this value).
-        if (this == null) {
-          throw new TypeError('"this" is null or not defined');
-        }
-
-        var o = Object(this);
-
-        // 2. Let len be ? ToLength(? Get(O, "length")).
-        var len = o.length >>> 0;
-
-        // 3. If len is 0, return false.
-        if (len === 0) {
-          return false;
-        }
-
-        // 4. Let n be ? ToInteger(fromIndex).
-        //    (If fromIndex is undefined, this step produces the value 0.)
-        var n = fromIndex | 0;
-
-        // 5. If n ≥ 0, then
-        //  a. Let k be n.
-        // 6. Else n < 0,
-        //  a. Let k be len + n.
-        //  b. If k < 0, let k be 0.
-        var k = Math.max(n >= 0 ? n : len - Math.abs(n), 0);
-
-        function sameValueZero(x, y) {
-          return (
-            x === y ||
-            (typeof x === 'number' &&
-              typeof y === 'number' &&
-              isNaN(x) &&
-              isNaN(y))
-          );
-        }
-
-        // 7. Repeat, while k < len
-        while (k < len) {
-          // a. Let elementK be the result of ? Get(O, ! ToString(k)).
-          // b. If SameValueZero(searchElement, elementK) is true, return true.
-          // c. Increase k by 1.
-          if (sameValueZero(o[k], searchElement)) {
-            return true;
-          }
-          k++;
-        }
-
-        // 8. Return false
-        return false;
-      }
-    });
+export function evalscriptoverridesToString(overridesObj) {
+  let encodedOverrides = '';
+  if (overridesObj.gainOverride) {
+    encodedOverrides += `gainOverride=${overridesObj.gainOverride};`;
+  }
+  if (overridesObj.gammaOverride) {
+    encodedOverrides += `gammaOverride=${overridesObj.gammaOverride};`;
   }
 
-  if (!Array.prototype.map) {
-    // prettier-ignore
-    Array.prototype.map = function(callback /*, thisArg*/) { // eslint-disable-line no-extend-native
-      var T, A, k;
-
-      if (this == null) {
-        throw new TypeError('this is null or not defined');
-      }
-
-      // 1. Let O be the result of calling ToObject passing the |this|
-      //    value as the argument.
-      var O = Object(this);
-
-      // 2. Let lenValue be the result of calling the Get internal
-      //    method of O with the argument "length".
-      // 3. Let len be ToUint32(lenValue).
-      var len = O.length >>> 0;
-
-      // 4. If IsCallable(callback) is false, throw a TypeError exception.
-      // See: http://es5.github.com/#x9.11
-      if (typeof callback !== 'function') {
-        throw new TypeError(callback + ' is not a function');
-      }
-
-      // 5. If thisArg was supplied, let T be thisArg; else let T be undefined.
-      if (arguments.length > 1) {
-        T = arguments[1];
-      }
-
-      // 6. Let A be a new array created as if by the expression new Array(len)
-      //    where Array is the standard built-in constructor with that name and
-      //    len is the value of len.
-      A = new Array(len);
-
-      // 7. Let k be 0
-      k = 0;
-
-      // 8. Repeat, while k < len
-      while (k < len) {
-        var kValue, mappedValue;
-        if (k in O) {
-          // i. Let kValue be the result of calling the Get internal
-          //    method of O with argument Pk.
-          kValue = O[k];
-
-          // ii. Let mappedValue be the result of calling the Call internal
-          //     method of callback with T as the this value and argument
-          //     list containing kValue, k, and O.
-          mappedValue = callback.call(T, kValue, k, O);
-
-          // For best browser support, use the following:
-          A[k] = mappedValue;
-        }
-        // d. Increase k by 1.
-        k++;
-      }
-
-      // 9. return A
-      return A;
-    };
+  if (checkIfFisLayer()) {
+    if (overridesObj.valueRangeOverride) {
+      encodedOverrides += `rangeOverrides=${JSON.stringify([overridesObj.valueRangeOverride])};`;
+    }
+  } else if (
+    overridesObj.redRangeOverride ||
+    overridesObj.greenRangeOverride ||
+    overridesObj.blueRangeOverride
+  ) {
+    let rgbRangeOverrides = [[0, 1], [0, 1], [0, 1]];
+    if (overridesObj.redRangeOverride) {
+      rgbRangeOverrides[0] = overridesObj.redRangeOverride;
+    }
+    if (overridesObj.greenRangeOverride) {
+      rgbRangeOverrides[1] = overridesObj.greenRangeOverride;
+    }
+    if (overridesObj.blueRangeOverride) {
+      rgbRangeOverrides[2] = overridesObj.blueRangeOverride;
+    }
+    encodedOverrides += `rangeOverrides=${JSON.stringify(rgbRangeOverrides)};`;
   }
 
-  if (!String.prototype.includes) {
-    // prettier-ignore
-    String.prototype.includes = function(search, start) { // eslint-disable-line no-extend-native
-      if (typeof start !== 'number') {
-        start = 0;
-      }
+  return encodedOverrides;
+}
 
-      if (start + search.length > this.length) {
-        return false;
-      } else {
-        return this.indexOf(search, start) !== -1;
-      }
-    };
-  }
-  /* Math.log10 polyfill for internet explorer */
-  if (!Math.log10) {
-    Math.log10 = function(x) {
-      return Math.log(x) * Math.LOG10E;
-    };
-  }
+// JUST FOR SENTINEL-2
+export function checkIfFisLayer() {
+  const isSelectedResult = !!Store.current.selectedResult;
+  const isCustomLayer = Store.current.selectedResult && isCustomPreset(Store.current.selectedResult.preset);
+  const isShadowLayerAvailable =
+    Store.current.selectedResult &&
+    !!getFisShadowLayer(Store.current.selectedResult.name, Store.current.selectedResult.preset);
+  const isFisAvailableOnDatasource = !!(
+    Store.current.selectedResult && Store.current.selectedResult.baseUrls.FIS
+  );
+
+  return isSelectedResult && isFisAvailableOnDatasource && (isShadowLayerAvailable || isCustomLayer);
+}
+
+export function checkIfKnownValueLayer() {
+  // NDVI
+  // moisture index
+  // NDWI
+  //
+
+  return null;
 }
